Module ink_lang_ir::ast

Types and abstractions for ink! definitions that require custom syntax.

Note

In general we try not to require any sort of custom non-standard Rust syntax.

At the time of this writing we currently only use this for the argument parsing of ink! config header #[ink(env_types = my::env_types::Types, etc...)] in order to be able to parse identifiers in name = value segments for the value part.

Re-exports

pub use self::attr_args::AttributeArgs;

pub use self::attr_args::MetaNameValue;

pub use self::attr_args::PathOrLit;

Modules

attr_args