Module ink_lang_ir::ir::attrs

Structs

AttributeArg
An ink! specific attribute argument.

InkAttribute
An ink! specific attribute.

Namespace
An ink! namespace applicable to a trait implementation block.

Enums

Attribute
Either an ink! specific attribute, or another uninterpreted attribute.

AttributeArgKind
An ink! specific attribute flag.

Traits

Attrs
Types implementing this trait can return a slice over their syn attributes.

Functions

contains_ink_attributes
Returns true if the given iterator yields at least one attribute of the form #[ink(..)] or #[ink].

err_non_hex
Returns an error to notify about non-hex digits at a position.

first_ink_attribute
Returns the first valid ink! attribute, if any.

invalid_selector_err_regex
Returns an error to notify about an invalid ink! selector.

partition_attributes
Partitions the given attributes into ink! specific and non-ink! specific attributes.

sanitize_attributes
Sanitizes the given attributes.