Module ink_lang_ir::ir::idents_lint

Modules

private

Traits

VisitBy
Makes sure to call the correct visitor function on the given visitor.

Functions

ensure_no_ink_identifiers
Returns Ok if there are no occurrences of identifiers starting with __ink_.