Struct ink_lang_ir::ir::idents_lint::private::IdentVisitor

pub struct IdentVisitor {
    errors: Vec<Error>,
}

Visitor to ensure that there are no identifiers starting with __ink_ as prefix.

Errors

If there are identifiers starting with __ink_ as prefix in the input. Will yield one combined error for all found encounters.

Fields
errors: Vec<Error>
Implementations

impl IdentVisitor

pub fn into_result(self) -> Result<(), Error>

Converts the visitor into the errors it found if any.

Returns Ok if it found no errors during visitation.