Module ink_lang_ir::ir::item_impl::callable

Utilities and helper routines that are useful for both ink! messages and ink! constructors.

Structs

CallableWithSelector
Wrapper for a callable that adds its composed selector.

InputsIter
Iterator over the input parameters of an ink! message or constructor.

Enums

CallableKind
The kind of externally callable smart contract entity.

Visibility
The visibility of an ink! message or constructor.

Traits

Callable
An ink! callable.

Functions

compose_selector
Returns the composed selector of the ink! callable.

ensure_callable_invariants
Ensures that common invariants of externally callable ink! entities are met.