Function ink_lang_ir::ir::item_impl::callable::compose_selector

pub fn compose_selector<C>(item_impl: &ItemImpl, callable: &C) -> Selector where
    C: Callable, 

Returns the composed selector of the ink! callable.

Composition takes into account the given ir::ItemImpl.

Details

Given

the callable's identifier i

the optionally set callable's selector s

the impl blocks trait path in case it implements a trait, P

the impl blocks optional user provided namespace S

Then the selector is composed in the following way:

If s is given we simply return s.

Otherwise if T is not None (trait impl block) we concatenate S, T and i with :: as separator if T refers to a full-path. If T refers to a relative path or is just an identifier we only take its last segment p (e.g. the trait's identifier) into consideration and use it instead of P in the above concatenation. In the following we refer to the resulting concatenation as C.

Now we take the BLAKE-2 hash of C which results in 32 bytes of output and take the first 4 bytes that are returned in order as the composed selector.

Examples

Overriding the composed selector

Given

impl MyStorage {
    #[ink(message, selector = "0xDEADBEEF")]
    fn my_message(&self) {}
}

... then the selector of my_message is simply 0xDEADBEEF since it overrides the composed selector.

Inherent implementation block

Given

impl MyStorage {
    #[ink(message)]
    fn my_message(&self) {}
}

... then the selector of my_message is composed such as:

BLAKE2("my_message".to_string().as_bytes())[0..4]

Trait implementation block

Given

impl MyTrait for MyStorage {
    #[ink(message)]
    fn my_message(&self) {}
}

... then the selector of my_message is composed such as:

BLAKE2("MyTrait::my_message".to_string().as_bytes())[0..4]

Using full path for trait

Given

impl ::my_full::long_path::MyTrait for MyStorage {
    #[ink(message)]
    fn my_message(&self) {}
}

... then the selector of my_message is composed such as:

BLAKE2("::my_full::long_path::MyTrait::my_message".to_string().as_bytes())[0..4]

Using a namespace

Given

#[ink(namespace = "my_namespace")]
impl MyTrait for MyStorage {
    #[ink(message)]
    fn my_message(&self) {}
}

... then the selector of my_message is composed such as:

BLAKE2("my_namespace::MyTrait::my_message".to_string().as_bytes())[0..4]

Note

All above examples work similarly for ink! constructors interchangeably.

Usage Recommendations

These recommendation mainly apply to trait implementation blocks:

The recommendation by the ink! team is to use the full-path approach wherever possible; OR import the trait and use only its identifier with an additional namespace if required to disambiguate selectors.

Try not to intermix the above recommendations.

Avoid directly setting the selector of an ink! message or constuctor. Only do this if nothing else helps and you need a very specific selector, e.g. in case of backwards compatibility.

Do not use the namespace unless required to disambiguate.