Function ink_lang_ir::ir::item_impl::callable::ensure_callable_invariants

pub(in ir::item_impl) fn ensure_callable_invariants(
    method_item: &ImplItemMethod, 
    kind: CallableKind
) -> Result<(), Error>

Ensures that common invariants of externally callable ink! entities are met.

Errors

In case any of the common externally callable invariants are not met:

This is true if the externally callable is:

generic

const (compile-time evaluatable)

async (async WebAssembly smart contract calling is not allowed)

unsafe (caller provided assertions not yet stable)

Furthermore this is true if the externally callable is defined for a non default ABI (e.g. extern "C") or doesn't have valid visibility.