Function ink_lang_ir::ir::utils::ensure_pub_visibility

pub fn ensure_pub_visibility(
    name: &str, 
    parent_span: Span, 
    vis: &Visibility
) -> Result<(), Error>

Ensures that the given visibility is pub and anotherwise returns an appropriate error.

Note

The name parameter is given to improve the resulting error message. It denotes the entity which cannot have non-public visibility.