Struct ink_lang_ir::Config

pub struct Config { /* fields omitted */ }

The ink! configuration.

Implementations

impl Config

pub fn env_types(&self) -> Path

Returns the environmental types definition if specified. Otherwise returns the default environmental types definition provided by ink!.

pub fn is_dynamic_storage_allocator_enabled(&self) -> bool

Returns true if the dynamic storage allocator facilities are enabled for the ink! smart contract, false otherwise.

If nothing has been specified returns the default which is false.

pub fn is_compile_as_dependency_enabled(&self) -> bool

Return true if this ink! smart contract shall always be compiled as if it was a dependency of another smart contract, returns false otherwise.

If nothing has been specified returns the default which is false.