Struct ink_lang_ir::ItemImpl

pub struct ItemImpl { /* private fields */ }

An ink! implementation block.

Note

This can be either an inherent implementation block that implements some constructors, messages or internal functions for the storage struct; OR it can be a trait implementation for the storage struct.

We try to support all fields that are supported by the underlying syn implementation for syn::ItemImpl even though they are not really required to represent ink!. This is done for consistency with syn.

Implementations

impl ItemImpl

pub fn attrs(&self) -> &[Attribute]

Returns all non-ink! specific attributes of the implementation block.

pub fn self_type(&self) -> &Type

Returns the Self type of the implementation block.

pub fn trait_path(&self) -> Option<&Path>

Returns the trait type path if this is a trait implementation block.

Returns None if this is an inherent implementation block.

pub fn trait_ident(&self) -> Option<&Ident>

Returns the trait identifier if this is a trait implementation block.

Returns None if this is an inherent implementation block.

pub fn namespace(&self) -> Option<&Namespace>

Returns the namespace of the implementation block if any has been provided.

pub fn iter_messages(&self) -> IterMessages<'_>

Returns an iterator yielding the ink! messages of the implementation block.

pub fn iter_constructors(&self) -> IterConstructors<'_>

Returns an iterator yielding the ink! messages of the implementation block.

pub fn items(&self) -> &[ImplItem]

Returns a slice over the shared references of the items of the impl.