Enum ink_lang_ir::ImplItem

pub enum ImplItem {
    Constructor(Constructor),
    Message(Message),
    Other(ImplItem),
}

An item within an ink! implementation block.

Can be either

an ink! ir::Constructor

an ink! ir::Message

or any other non-ink! item.

Note

Based on syn::ImplItem with special variants for ink! impl items.

Variants

Constructor(Constructor)

A #[ink(constructor)] marked inherent function.

Message(Message)

A #[ink(message)] marked method.

Other(ImplItem)

Any other implementation block item.

Implementations

impl ImplItem

pub fn is_message(&self) -> bool

Returns true if the impl block item is an ink! message.

pub fn filter_map_message(&self) -> Option<&Message>

Returns Some if self is an ink! message.

Otherwise, returns None.

pub fn is_constructor(&self) -> bool

Returns true if the impl block item is an ink! message.

pub fn filter_map_constructor(&self) -> Option<&Constructor>

Returns Some if self is an ink! constructor.

Otherwise, returns None.

pub fn is_other_item(&self) -> bool

Returns true if the impl block item is a non ink! specific item.

pub fn filter_map_other_item(&self) -> Option<&ImplItem>

Returns Some if self is a not an ink! specific item.

Otherwise, returns None.